+++ ProtoCourse.WebApi/Controllers/LessonsController.cs
@@ -5,6 +5,7 @@ using ProtoCourse.Core.Models.Course;
 using ProtoCourse.Core.Models;
 using ProtoCourse.Core.Repository;
 using ProtoCourse.Core.Models.Lesson;
+using ProtoCourse.Core.Exceptions;
 
 namespace ProtoCourse.WebApi.Controllers;
 
@@ -14,42 +15,33 @@ public class LessonsController : ControllerBase
 {
     private readonly ILogger<LessonsController> _logger;
     private readonly ILessonsRepository _lessonsRepository;
+    private readonly IUserRepository _userRepository;
 
-    public LessonsController(ILessonsRepository lessonsRepository, ILogger<LessonsController> logger)
-        => (_logger, _lessonsRepository) = (logger, lessonsRepository);
+    public LessonsController(ILessonsRepository lessonsRepository, ILogger<LessonsController> logger, IUserRepository userRepository)
+        => (_logger, _lessonsRepository, _userRepository) = (logger, lessonsRepository, userRepository);
 
-    [HttpGet("get-all")]
-    public async Task<ActionResult<IEnumerable<LessonDto>>> GetLessons()
-    {
-        var courses = await _lessonsRepository.GetAllAsync();
-        return Ok(courses);
-    }
-
-    [HttpGet]
-    public async Task<ActionResult<PagedResult<LessonDto>>> GetPagedLessons([FromQuery] QueryParameters queryParameters)
-    {
-        var courses = await _lessonsRepository.GetAllAsync<LessonDto>(queryParameters);
-        return Ok(courses);
-    }
-
-    [HttpGet("{id:guid}")]
-    public async Task<ActionResult<LessonDto>> GetLesson(Guid id)
-    {
-        var course = await _lessonsRepository.GetAsync(id);
-        return Ok(course);
-    }
+    #region AuthorAccess
 
     [HttpPost]
     [Authorize]
-    public async Task<ActionResult> CreateLesson(CreateCourseDto createLessonDto)
+    public async Task<ActionResult> CreateLesson(CreateLessonDto createLessonDto)
     {
-        var course = await _lessonsRepository.AddAsync<CreateCourseDto, GetCourseDto>(createLessonDto);
-        return CreatedAtAction(nameof(GetLesson), new { id = course.Id }, course);
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+
+        if (!await _userRepository.IsUserAuthorOfCourse(userId, createLessonDto.CourseId)) throw new ForbiddenException();
+
+        var lessonDto = await _lessonsRepository.AddAsync<CreateLessonDto, LessonDto>(createLessonDto);
+        return CreatedAtAction(nameof(GetLesson), new { id = lessonDto.Id }, lessonDto);
     }
 
     [HttpPut("{id:guid}")]
+    [Authorize]
     public async Task<ActionResult> UpdateLesson(Guid id, UpdateLessonDto updateLessonDto)
     {
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+
+        if (!await _userRepository.IsUserAuthorOfCourse(userId, id)) throw new ForbiddenException();
+
         await _lessonsRepository.UpdateAsync(id, updateLessonDto);
         return Ok();
     }
@@ -58,7 +50,56 @@ public class LessonsController : ControllerBase
     [Authorize]
     public async Task<ActionResult> DeleteLesson(Guid id)
     {
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+        var courseId = (await _lessonsRepository.GetAsync(id)).CourseId;
+
+        if (!await _userRepository.IsUserAuthorOfCourse(userId, courseId)) throw new ForbiddenException();
+
         await _lessonsRepository.DeleteAsync(id);
         return NoContent();
     }
+
+    #endregion
+
+    #region StudentAccess
+
+    [HttpGet("course/{id:guid}/get-all")]
+    [Authorize]
+    public async Task<ActionResult<IEnumerable<LessonDto>>> GetLessons(Guid id)
+    {
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+
+        if (!await _userRepository.IsUserStudentOfCourse(userId, id)) throw new ForbiddenException();
+
+        var courses = await _lessonsRepository.GetAllLessonsAsync<LessonDto>(id);
+        return Ok(courses);
+    }
+
+    [HttpGet("course/{id:guid}")]
+    [Authorize]
+    public async Task<ActionResult<PagedResult<LessonDto>>> GetPagedLessons(Guid id, [FromQuery] QueryParameters queryParameters)
+    {
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+
+        if (!await _userRepository.IsUserStudentOfCourse(userId, id)) throw new ForbiddenException();
+
+        var courses = await _lessonsRepository.GetAllAsync<LessonDto>(queryParameters);
+        return Ok(courses);
+    }
+
+    [HttpGet("{id:guid}")]
+    [Authorize]
+    public async Task<ActionResult<LessonDto>> GetLesson(Guid id)
+    {
+        var lessonDto = await _lessonsRepository.GetAsync<LessonDto>(id);
+        var userId = HttpContext.User.FindFirst("uid")!.Value;
+
+        if (!await _userRepository.IsUserStudentOfCourse(userId, lessonDto.CourseId) ||
+            !await _userRepository.IsUserAuthorOfCourse(userId, lessonDto.CourseId))
+            throw new ForbiddenException();
+
+        return Ok(lessonDto);
+    }
+
+    #endregion
 }
